# [TS/JS] Create byte vectors