@razvanalex 2023-12-12
It seems the failure of labeler action is related to [this](https://github.com/actions/labeler/issues/712).